[release-v1.59] Catch up with non-logical main changes for ease of maintenance (#3227)
* Double number of e2e parallel nodes from 3 to 6 (#3156)

We have some feedback from different platform testing that this
can work tests (not resource) wise. Let's give it a try here as well


* Support IPv6 in controller GetMetricsURL (#3165)

Joining hostname+port with string concatenation leads to wrong URLs
when the hostame is an IPv6:

HOST    PORT   Sprintf    CORRECT
::1     8000   ::1:8000   [::1]:8000

To avoid this issue, it's best to use net.JoinHostPort. I added a test
that verifies this fix.

This type of issue can be discovered running the following command:

    golangci-lint run ./... --enable nosprintfhostport


* Enable misspell linter and fix spelling errors (#3164)

* Add misspell to list of linters


* Fix spelling errors


---------


* Remove openshift dep on api (#3172)

Less deps the better!  Could be problematic for projects importing CDI and openshift.


* Enable unconvert linter (#3171)

* Enable unconvert linter

This linter's doc describes it as:

   The unconvert program analyzes Go packages to identify unnecessary
   type conversions; i.e., expressions T(x) where x already has type T.


* Unrestrict the number of linter warnings

It is best to show all warnings at once than to reveal them piece-meal,
particularly in CI where the feedback loop can be a bit slow.

By default, linters may only print the same message three times
(https://golangci-lint.run/usage/configuration/#issues-configuration)
The unconvert linter always prints the same message, so it specially
affected by this setting.


* Remove redundant type conversions


---------


* Enable linters checking error handling (#3177)

* Enable errorlint linter

Per the readme:

> go-errorlint is a source code linter for Go software that can be used
to find code that will cause problems with the error wrapping scheme
introduced in Go 1.13.

https://github.com/polyfloyd/go-errorlint

Essentially, it checks that you properly use errors.Is and errors.As

It also forces you to use %w, but that is a bit too much. Using %v is
fine when you don't want to leak the internal error types of the
package. This is why I disabled this setting.


* Replace error comparisons with errors.Is


* Replace error type assertions with errors.As


* Enable errname linter

Per the readme:
> Checks that sentinel errors are prefixed with the Err and error
> types are suffixed with the Error.

https://github.com/Antonboom/errname

Not a big deal, but helps keep the naming consistent.


* Fix errname linter warnings


---------


* Run bazelisk run //robots/cmd/uploader:uploader -- -workspace /home/prow/go/src/github.com/kubevirt/project-infra/../containerized-data-importer/WORKSPACE -dry-run=false (#3199)


* CVE 2024-24786 fix: Bump google.golang.org/protobuf from 1.31.0 to 1.33.0 (#3195)

* Upgrade google.golang.org/protobuf

This solves CVE-2024-24786

https://www.cve.org/CVERecord?id=CVE-2024-24786


* Update checksum and vendoring


---------


* Enable durationcheck linter (#3176)

* Enable durationcheck linter

> Check for two durations multiplied together.

https://github.com/charithe/durationcheck


* Fix misuse of time.Duration

I also had to rename the variable because go-statickcheck complains
about a time.Duration having the suffix 'Sec'.


---------


* Enable nakedret linter (#3178)

* Enable nakedret linter

> Checks that functions with naked returns are not longer than a maximum
> size (can be zero).

https://github.com/alexkohler/nakedret

Relevant from the GO style guide:
> Naked returns are okay if the function is a handful of lines. Once
> it’s a medium sized function, be explicit with your return values.

https://go.dev/wiki/CodeReviewComments#named-result-parameters

I think a naked return is never easier to read than a regular return,
so I set the linter to always complain about it. Disagreements are
welcome.


* Refactor functions with naked returns


---------


* Clean up leftover snapshot sources from DataImportCron tests (#3123)

* Clean up leftover snapshot sources from DataImportCron tests

The dataimportcron tests may affect existing crons in the environment (in addition to the ones deployed by testing)
so we are better off cleaning up after ourselves.


* Add watch for volumesnapshot delete

Although we don't support seamless transition from volumesnapshot->pvc sources
(we hope you stay on snapshots if they scale better for your storage)
this will still be needed in most cases when someone switches and manually deletes snapshots.


---------


* Enable autoformatting linters (#3179)

* Enable gofmt linter

From the docs:

> Gofmt checks whether code was gofmt-ed. By default this tool runs with
> -s option to check for code simplification.

https://golangci-lint.run/usage/linters/#gofmt


* Run gomft on the project

Ran this command after adding the gofmt linter:

	golangci-lint run ./... --fix


* Enable whitespace linter

From the docs:
> Whitespace is a linter that checks for unnecessary newlines at the
> start and end of functions, if, for, etc.

https://golangci-lint.run/usage/linters/#whitespace


* Run whitespace on the project

Ran this command after adding the whitespace linter:

	golangci-lint run ./... --fix


* Enable GCI linter

Per the docs:

> Gci controls Go package import order and makes it always deterministic.

https://golangci-lint.run/usage/linters/#gci

NOTE: I noticed that many files separate their imports in a particular
way, so I set the linter to enforce this standard.


* Run GCI on the project

Ran this command after adding the GCI linter:

	golangci-lint run ./... --fix


---------


* Enable dupword linter (#3175)

* Enable dupword linter

Per the readme:
> A linter that checks for duplicate words in the source code (usually miswritten)

https://github.com/Abirdcfly/dupword
https://golangci-lint.run/usage/linters/#dupword


* Remove duplicate words in comments and strings


---------


* Change clone strategy and cron sources format for LVMS (#3196)

Following some investigation we are now confident this tuning is benefical for LVMS:
- CSI clone is effectively implemented the same as snapshotting because both already use lvm2 thinly provisioned snapshots under the hood
- Snapshots being COW, it makes total sense to store cron sources in the snapshot format instead of PVC:
```
A snapshot of a thin logical volume also creates a thin logical volume.
This consumes no data space until a COW operation is required, or until the snapshot itself is written.
```


* Set default clone strategy & cron format for trident to snapshot (#3209)

There's a bug in the trident CSI driver that causes a snapshot
to be left over following each CSI clone:
https://github.com/NetApp/trident/issues/901
This is becoming a problem rapidly once one hits the snapshot limit per volume (golden image):
https://kb.netapp.com/onprem/ontap/os/Maximum_number_of_snapshots_supported_by_ONTAP


---------

+++ pkg/controller/datavolume/import-controller_test.go
@@ -29,6 +29,7 @@ import (
 	. "github.com/onsi/gomega"
 
 	snapshotv1 "github.com/kubernetes-csi/external-snapshotter/client/v6/apis/volumesnapshot/v1"
+
 	corev1 "k8s.io/api/core/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	extv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
@@ -40,6 +41,7 @@ import (
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/tools/record"
 	"k8s.io/utils/ptr"
+
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
@@ -1700,7 +1702,7 @@ var _ = Describe("All DataVolume Tests", func() {
 		largeOverhead   = float64(0.75)
 	)
 	DescribeTable("GetRequiredSpace should return properly enlarged sizes,", func(imageSize int64, overhead float64) {
-		for testedSize := int64(imageSize - 1024); testedSize < imageSize+1024; testedSize++ {
+		for testedSize := imageSize - 1024; testedSize < imageSize+1024; testedSize++ {
 			alignedImageSpace := imageSize
 			if testedSize > imageSize {
 				alignedImageSpace = imageSize + Mi
